Make pythonCall with builtin types/functions work correctly
Previously if the callee of a pythonCall match is a built-in type
or function (pythonBuiltinType, e.g. dict, str), the syntax rule
`pythonCall` was not working as expected. This is because such
built-in keywords (defined in python-mode) do not have 'contained'
property, which is out of our control.

As described in https://vi.stackexchange.com/questions/18408/ ,
we can re-define, for instance, a syntax group `pythonCallBuiltinType`
by mimicking `pythonBuiltinType` and make it able to be contained.